Renamed core.py to export.py as it make more sense
